+++ src/routes/(auth)/weather/components/ForecastItem.svelte
