+++ src/routes/(auth)/weather/components/ResultsCard.svelte
